jsonrpc: convert currency-info test to forge
